% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pred5mc.R
\name{Pred5mc}
\alias{Pred5mc}
\title{Pred5mc}
\usage{
Pred5mc(Fastafile)
}
\arguments{
\item{Fastafile}{Sequence file (.fasta format)}
}
\value{
MethStatus: Sequences with their methylation state (methylated or non-methylated)
}
\description{
Prediction of sequences with 5mc sites.
}
\examples{
\donttest{
library(GB5mcPred)
data<-system.file("exdata/test.fasta", package = "GB5mcPred")
pred<-Pred5mc(Fastafile=data)
}
}
\references{
Chen, W., Lv, H., Nie, F., & Lin, H. (2019). i6mA-Pred: identifying DNA N6-methyladenine 	sites in the rice genome. Bioinformatics, 35(16), 2796-2800.
}
